+++ src/browser/settings.jsx
@@ -47,6 +47,13 @@ var SettingsPage = React.createClass({
         });
       });
     }
+
+    if (process.platform === 'darwin') {
+      var currentWindow = remote.getCurrentWindow();
+      this.setState({
+        trayWasVisible: currentWindow.trayWasVisible
+      });
+    }
   },
   handleTeamsChange: function(teams) {
     this.setState({
@@ -62,6 +69,8 @@ var SettingsPage = React.createClass({
       trayIconTheme: this.state.trayIconTheme,
       disablewebsecurity: this.state.disablewebsecurity,
       version: settings.version,
+      minimizeToTray: this.state.minimizeToTray,
+      toggleWindowOnTrayIconClick: this.state.toggleWindowOnTrayIconClick,
       notifications: {
         flashWindow: this.state.notifications.flashWindow
       }
@@ -101,9 +110,16 @@ var SettingsPage = React.createClass({
     });
   },
   handleChangeShowTrayIcon: function() {
+    var shouldShowTrayIcon = this.refs.showTrayIcon.getChecked();
     this.setState({
-      showTrayIcon: this.refs.showTrayIcon.getChecked()
+      showTrayIcon: shouldShowTrayIcon
     });
+
+    if (process.platform === 'darwin' && !shouldShowTrayIcon) {
+      this.setState({
+        minimizeToTray: false
+      });
+    }
   },
   handleChangeTrayIconTheme: function() {
     this.setState({
@@ -113,6 +129,19 @@ var SettingsPage = React.createClass({
   handleChangeAutoStart: function() {
     this.setState({
       autostart: this.refs.autostart.getChecked()
+    });
+  },
+  handleChangeMinimizeToTray: function() {
+    var shouldMinimizeToTray = (process.platform !== 'darwin' || this.refs.showTrayIcon.getChecked())
+    && this.refs.minimizeToTray.getChecked();
+
+    this.setState({
+      minimizeToTray: shouldMinimizeToTray
+    });
+  },
+  handleChangeToggleWindowOnTrayIconClick: function() {
+    this.setState({
+      toggleWindowOnTrayIconClick: this.refs.toggleWindowOnTrayIconClick.getChecked()
     });
   },
   toggleShowTeamForm: function() {
@@ -158,6 +187,17 @@ var SettingsPage = React.createClass({
       options.push(<Input key="inputAutoStart" id="inputAutoStart" ref="autostart" type="checkbox" label="Start app on login." checked={ this.state.autostart } onChange={ this.handleChangeAutoStart }
                    />);
     }
+
+    if (process.platform === 'darwin') {
+      options.push(<Input key="inputMinimizeToTray" id="inputMinimizeToTray" ref="minimizeToTray" type="checkbox" label={ this.state.trayWasVisible || !this.state.showTrayIcon ? "Leave app running in notification area when the window is closed" : "Leave app running in notification area when the window is closed (available on next restart)" } disabled={ !this.state.showTrayIcon || !this.state.trayWasVisible } checked={ this.state.minimizeToTray }
+                     onChange={ this.handleChangeMinimizeToTray } />);
+    }
+
+    if (process.platform === 'win32') {
+      options.push(<Input key="inputToggleWindowOnTrayIconClick" id="inputToggleWindowOnTrayIconClick" ref="toggleWindowOnTrayIconClick" type="checkbox" label="Toggle window visibility when clicking on the tray icon."
+                     checked={ this.state.toggleWindowOnTrayIconClick } onChange={ this.handleChangeToggleWindowOnTrayIconClick } />);
+    }
+
     var options_row = (options.length > 0) ? (
       <Row>
         <Col md={ 12 }>